pyproject.toml
[project]
name = "thundechef-api"
version = "0.1.0"
description = "Thunderchef backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[standard]>=0.117.1",
    "openai[aiohttp]>=1.109.1",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.36.0",
]
